Version 1.2.0
- Added Brightness/Contrast widget, toggled by F3, to change brightness/contrast in any mouse mode and
provide visual information about current settings.
- Added annotation source tracking feature. Annotations are associated with a metadata object containing
information about their origin, whether it is auto-generated by certain method with certain parameters, or
added/modified by a human on certain date. The metadata is displayed in the status line (at bottom right
of the main screen) when hovering mouse over an annotation. In addition, there is a dialog (which is open via
menu View -> Annotation Sources... or F6), which allows to gray out annotations associated with selected sources.
The dialog also has a "Comment" box to add user-defined comments on current edits.
- Added "Export Annotation Stats..." item to the "File" menu. It exports annotation tracker statistics (like
who and when created/deleted/modified how many annotations).
- Added possibility to open 3D image stacks. They are marked with a different icon in the list of open files,
when the user double-clicks on such entry, a list of frames is displayed, so the user can browse across
separate frames and select ones intended for automatic processing (auto-detection).